+++ src/main/java/com/teamwork/project/projects/mapper/TaskMapper.java
@@ -3,13 +3,13 @@ package com.teamwork.project.projects.mapper;
 import com.teamwork.project.projects.domain.Task;
 
 public interface TaskMapper {
-    int deleteByPrimaryKey(Integer taskId);
+    int deleteByPrimaryKey(Long taskId);
 
     int insert(Task record);
 
     int insertSelective(Task record);
 
-    Task selectByPrimaryKey(Integer taskId);
+    Task selectByPrimaryKey(Long taskId);
 
     int updateByPrimaryKeySelective(Task record);
